requirements.txt
fastmcp>=2.6.1
neo4j>=5.26.0
python-dotenv>=1.0.0
uvicorn>=0.34.0
starlette>=0.47.0
pydantic>=2.11.0
pydantic-settings>=2.9.0 